requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
sqlalchemy
pandas
azure-storage-blob
openpyxl
azure-identity
azure-keyvault-secrets